+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Made `embedded-hal` and `embedded-io-nb` trait implementations optional, behind corresponding
   feature flags.
 - Added methods to read, mask and clear interrupts.
+- Added methods to set FIFO level of RX/TX interrupts
 
 ## 0.1.0